+++ native/explorer/src/lib.rs
@@ -13,7 +13,6 @@ mod error;
 mod lazyframe;
 mod series;
 
-use dataframe::*;
 pub use datatypes::{
     ExAnyValue, ExDataFrame, ExDataFrameRef, ExLazyFrame, ExLazyFrameRef, ExSeries, ExSeriesRef,
 };
@@ -39,52 +38,34 @@ mod atoms {
 rustler::init!(
     "Elixir.Explorer.PolarsBackend.Native",
     [
-        // dataframe
-        df_as_str,
-        df_column,
-        df_columns,
-        df_drop,
-        df_drop_duplicates,
-        df_drop_nulls,
-        df_dtypes,
-        df_fill_none,
-        df_filter,
-        df_from_map_rows,
-        df_from_keyword_rows,
-        df_get_columns,
-        df_groups,
-        df_groupby_agg,
-        df_head,
-        df_height,
-        df_hstack,
-        df_join,
-        df_melt,
-        df_new,
-        df_pivot_wider,
-        df_read_csv,
-        df_read_parquet,
-        df_select,
-        df_select_at_idx,
-        df_set_column_names,
-        df_shape,
-        df_slice,
-        df_sort,
-        df_tail,
-        df_take,
-        df_to_csv,
-        df_to_csv_file,
-        df_to_dummies,
-        df_vstack,
-        df_width,
-        df_with_column,
-        df_write_parquet,
         // lazyframe
+        lf_as_str,
         lf_collect,
+        lf_column_dtypes,
+        lf_column_names,
         lf_describe_plan,
         lf_drop,
+        lf_drop_nulls,
         lf_fetch,
+        lf_from_keyword_rows,
+        lf_from_map_rows,
+        lf_get_column,
+        lf_get_columns,
+        lf_height,
+        lf_read_csv,
+        lf_read_parquet,
         lf_select,
+        lf_shape,
         lf_tail,
+        lf_to_csv_string,
+        lf_width,
+        lf_write_csv,
+        lf_write_parquet,
+        lf_new,
+        lf_distinct,
+        lf_melt,
+        lf_sort,
+        lf_drop_nulls,
         // series
         s_add,
         s_and,
